Extract functions to another module
